requirements.txt
# AbemaTV Episode Extractor Requirements

# Core dependencies
PyYAML==6.0.3